requirements.txt
six==1.10.0
setuptools==41.0.1
docopt==0.6.2
alpaca_trade_api==0.29
pandas==0.24.1
pendulum==2.0.5
PyYAML==5.1.2
